renamed injection config file
