+++ lib/injection/injection.dart
@@ -1,7 +1,7 @@
 import 'package:get_it/get_it.dart';
 import 'package:injectable/injectable.dart';
 
-import 'injection.config.dart';
+import 'injection_config.dart';
 
 final GetIt getIt = GetIt.instance;
